added documentation
