--- src/BookingType.js
@@ -1,4 +1,0 @@
-export const BookingType = {
-  'PER_DAY': 'D',
-  'PER_KM': 'K'
-};
