+++ src/Models/FuelPurchase.js
@@ -1,5 +1,12 @@
+/**
+ * FuelPurchase.js
+ */
 import moment from "moment";
 
+/**
+ * FuelPurchase model class
+ * @class FuelPurchase
+ */
 export class FuelPurchase {
   _id;
   _bookingID;
@@ -8,6 +15,15 @@ export class FuelPurchase {
   _fuelQuantity;
   _updatedAt;
 
+  /**
+   * Creates a new FuelPurchase
+   * @param {string} bookingID - ID of the booking associated with this fuel purchase
+   * @param {number} fuelQuantity - amount of fuel of this purchase (in litres)
+   * @param {number} fuelPrice - price per litre of this fuel purchase
+   * @param {string} id - ID of this fuel purchase
+   * @param {string} createdAt - timestamp generated when this fuel purchase is created
+   * @param {string|null} updatedAt - timestamp generated when this fuel purchase is updated
+   */
   constructor(bookingID, fuelQuantity, fuelPrice, id = require('uuid/v4')(), createdAt = moment().format('DD/MM/YYYY hh:mm:ss A'), updatedAt = null) {
     this._id = id;
     this._bookingID = bookingID;
