+++ src/components/AddJourney/AddJourney.js
@@ -1,3 +1,6 @@
+/**
+ * AddJourney.js
+ */
 import React from 'react';
 import {AppConsumer, AppContext} from "../../AppContext/AppContext";
 import {Accordion, Button, Card, Col, Container, ListGroup, Row} from "react-bootstrap";
@@ -7,6 +10,9 @@ import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
 import {faPlus} from "@fortawesome/free-solid-svg-icons/faPlus";
 import {Link} from "react-router-dom";
 
+/**
+ * AddJourney component - navigates to AddJourneyForm once a booking is selected
+ */
 export class AddJourney extends React.Component {
   constructor(props) {
 	super(props);
