+++ src/components/LoadingSpinner/LoadingSpinner.js
@@ -1,6 +1,14 @@
+/**
+ * LoadingSpinner.js
+ */
 import React from 'react';
 import { Spinner } from 'react-bootstrap';
 
+/**
+ * LoadingSpinner component - renders a loading spinner, with a fallback for screen readers
+ * @returns {*}
+ * @constructor
+ */
 export const LoadingSpinner = () => (
   <Spinner
 	animation="border"
