+++ src/components/Navigation/Navigation.js
@@ -1,9 +1,15 @@
+/**
+ * Navigation.js
+ */
 import React from 'react';
 import {Navbar, Nav, NavDropdown} from 'react-bootstrap';
 import {Link} from 'react-router-dom';
 import NavItem from "react-bootstrap/NavItem";
 import {SearchVehicle} from "./SearchVehicle";
 
+/**
+ * Navigation component - renders a navbar to navigate through the app
+ */
 export class Navigation extends React.Component {
   render() {
 	return (
