+++ src/components/Notification/Notification.js
@@ -1,6 +1,21 @@
+/**
+ * Notification.js
+ */
 import React, {useState} from 'react';
 import {Toast} from "react-bootstrap";
 
+/**
+ * @typedef {Object} NotificationProps
+ * @property {boolean} display - determines whether a notification should be shown
+ * @property {string} message - notification message
+ */
+/**
+ * Notification component - renders notifications as a Toast, which is dismissible and auto hides
+ * in 3 seconds
+ * @param {NotificationProps} props
+ * @returns {*}
+ * @constructor
+ */
 export const Notification = props => {
   const [show, setShow] = useState(true);
