+++ src/index.js
@@ -1,3 +1,6 @@
+/**
+ * index.js
+ */
 import React from 'react';
 import ReactDOM from 'react-dom';
 import {App} from './components/App/App';
